Split island into modules
