+++ northstar/src/runtime/island/utils.rs
@@ -1,0 +1,29 @@
+// Copyright (c) 2021 ESRLabs
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+
+use std::path::{Path, PathBuf};
+
+pub(super) trait PathExt {
+    fn join_strip<T: AsRef<Path>>(&self, w: T) -> PathBuf;
+}
+
+impl PathExt for Path {
+    fn join_strip<T: AsRef<Path>>(&self, w: T) -> PathBuf {
+        if w.as_ref().starts_with("/") {
+            self.join(w.as_ref().strip_prefix("/").unwrap())
+        } else {
+            self.join(w)
+        }
+    }
+}
